+++ src/common/universalelectricity/extend/IElectricityStorage.java
@@ -10,15 +10,15 @@ public interface IElectricityStorage
 	/**
 	 * Returns the amount of watt hours this unit has stored.
 	 */
-	public float getWattHours(Object... data);
+	public double getWattHours(Object... data);
 	
 	/**
 	 * Sets the amount of watt hours this unit has stored.
 	 */
-	public void setWattHours(float wattHours, Object... data);
+	public void setWattHours(double wattHours, Object... data);
 	
 	/**
 	 * Gets the maximum amount of watt hours this unit can store.
 	 */
-	public float getMaxWattHours();
+	public double getMaxWattHours();
 }
